feat: use makeQueries for byText
